+++ HISTORY.rst
@@ -42,7 +42,7 @@ History
 * external import fix
 
 0.3.5 (2017-12-27)
-* GenericIPAddressField, MACAdressField included
+* GenericIPAddressField, MACAddressField included
 
 0.3.4 (2017-12-27)
 * ArrayField, TextField, also allow db_index for any kind of fields